add changes for db

+++ api/src/main/java/tpo/uade/api/controller/ProductController.java
@@ -39,7 +39,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ProductDto> getProductById(@PathVariable Long id) {
+    public ResponseEntity<ProductDto> getProductById(@PathVariable int id) {
         return ResponseEntity.ok(productService.getById(id));
     }
